Refactor App component to use CoffeeRecipe and simplify state management; enhance RecipeList with Header and Footer components
