+++ src/components/RecipeList.tsx
@@ -5,6 +5,9 @@ import { useNavigate } from 'react-router-dom';
 import { CoffeeRecipe } from '../types';
 import { useSettings } from '../context/SettingsContext';
 import { translations } from '../translations';
+import Header from './Header';
+import Footer from './Footer';
+import Container from '@mui/material/Container';
 
 interface RecipeListProps {
   recipes: CoffeeRecipe[];
@@ -21,40 +24,51 @@ const RecipeList: React.FC<RecipeListProps> = ({ recipes }) => {
   };
 
   return (
-    <Grid container spacing={3} sx={{ p: 2 }}>
-      {recipes.map((recipe) => (
-        <Grid size={{xs: 12, sm: 6}} key={recipe.id}>
-          <Card
-            sx={{
-              height: '100%',
-              display: 'flex',
-              flexDirection: 'column',
-              cursor: 'pointer',
-              '&:hover': {
-                transform: 'scale(1.02)',
-                transition: 'transform 0.2s ease-in-out',
-              },
-            }}
-            onClick={() => navigate(`/${language}/recipes/featured/${recipe.id}`)}
-          >
-            <CardMedia
-              component="img"
-              height="180"
-              image={getYouTubeThumbnail(recipe.youTubeEmbedUrl)}
-              alt={recipe.name[language]}
-            />
-            <CardContent>
-              <Typography gutterBottom variant="h5" component="div">
-                {recipe.name[language]}
-              </Typography>
-              <Typography variant="body2" color="text.secondary">
-                {recipe.description[language]}
-              </Typography>
-            </CardContent>
-          </Card>
-        </Grid>
-      ))}
-    </Grid>
+    <Container maxWidth="sm" sx={{
+      bgcolor: 'background.default',
+      color: 'text.primary',
+      minHeight: '100vh',
+      py: 2,
+      width: '100%',
+    }}>
+      <Header language={language} t={t} />
+
+      <Grid container spacing={3} sx={{ p: 2, mb: 2 }}>
+        {recipes.map((recipe) => (
+          <Grid size={{ xs: 12, sm: 6 }} key={recipe.id}>
+            <Card
+              sx={{
+                height: '100%',
+                display: 'flex',
+                flexDirection: 'column',
+                cursor: 'pointer',
+                '&:hover': {
+                  transform: 'scale(1.02)',
+                  transition: 'transform 0.2s ease-in-out',
+                },
+              }}
+              onClick={() => navigate(`/${language}/recipes/featured/${recipe.id}`)}
+            >
+              <CardMedia
+                component="img"
+                height="180"
+                image={getYouTubeThumbnail(recipe.youTubeEmbedUrl)}
+                alt={recipe.name[language]}
+              />
+              <CardContent>
+                <Typography gutterBottom variant="h5" component="div">
+                  {recipe.name[language]}
+                </Typography>
+                <Typography variant="body2" color="text.secondary">
+                  {recipe.description[language]}
+                </Typography>
+              </CardContent>
+            </Card>
+          </Grid>
+        ))}
+      </Grid>
+      <Footer t={t} />
+    </Container>
   );
 };
